feat: 좌석 홀드/결제확정 유즈케이스 추가
